Refactor data handling to class Storage
