+++ src/main/java/TaskList.java
@@ -1,0 +1,10 @@
+import java.util.ArrayList;
+import java.util.List;
+
+public class TaskList {
+    static List<Task> tasks = new ArrayList<>();
+
+    static String getList() {
+        return "";
+    }
+}
